operator overloading
This C++ code defines a class Increase with an integer member variable value, which is incremented by 5 each time the unary prefix increment operator ++ is applied to an object of this class. In the main() function, an object of Increase is created, incremented using the overloaded ++ operator, and its value is displayed.