Create Friend_function
This C++ code defines a class MyClass with a private member data. It also declares a friend function friendFunction that can access private members of MyClass. In the main() function, an object of MyClass is created, and the friend function is called to print the value of the private member data.