Create Friend person
This C++ code defines a Person class with private members name and age, and a constructor to initialize them. It also declares a friend function displayAge which can access the private members of Person. In the main() function, an object of Person is created with the name "Alice" and age 30, and then displayAge is called to print out the person's name and age.