Create division_operation
This code defines a class Divider with a member function divide to perform division. In the main() function, an object of the Divider class is created. The user is prompted to enter the dividend and divisor, and the divide() function is called to perform the division. If the divisor is zero, an error message is displayed, and the function returns 0. Otherwise, the result of the division is printed.